SAMWriter::from_bam_with_filter removed
